+++ Bluejay/Bluejay/FixedWidth.swift
@@ -1,0 +1,12 @@
+//
+//  FixedWidth.swift
+//  Bluejay
+//
+//  Created by Jeremy Chiang on 2019-03-11.
+//  Copyright © 2019 Steamclock Software. All rights reserved.
+//
+
+import Foundation
+
+/// Marker protocol to help enforce extraction only on primitives or fixed width types.
+public protocol FixedWidth {}